pyproject.toml
[project]
name = "event-analytics-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "cachetools>=6.2.4",
    "dishka>=1.7.2",
    "fastapi[standard]>=0.128.0",
    "granian>=2.6.0",
    "httpx>=0.28.1",
    "orjson>=3.11.5",
    "structlog>=25.5.0",
    "uuid6>=2025.0.1",
    "valkey>=6.1.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.10",
    "testcontainers[postgres]>=4.14.0",
    "types-cachetools>=6.2.0.20251022",
]

[tool.mypy]
python_version = "3.13"
strict = true
mypy_path = "src"
ignore_missing_imports = false
disallow_untyped_defs = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = ["asyncpg.*", "granian.*"]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = ["tests"]
select = [
    "F",    # Pyflakes
    "E",    # Pydocstyle error
    "W",    # Pydocstyle warnings
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "ANN",  # flake8-annotations
    "S",    # flake8-bandit
    "B",    # flake8-bugbear
    "C901", # mccabe complexity
    "SIM",  # flake8-simplify
    "T20",  # flake8-print
    "RUF",  # ruff specific rules
]
ignore = [
    "D203", # 1 blank line required before class docstring.
    "D213", # Multi-line docstring summary should start at the second line.
]

[tool.ruff.isort]
lines-after-imports = 2

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.per-file-ignores]
"src/tests/*" = [
    "S101", # access for assert in tests
    "ANN",  # Type annotations.
    "S105", # hardcode passwords in tests
]
"__init__.py" = [
    "F401", # Unused imports.
]

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"
pythonpath = "src"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
